Added cropping & attempt at making laswriter work
Filtered the points in lsPts, which covers the full tile, to filteredPts using a 500x500m bounding box. 

Also tried to write these points back to a LAS file with the function write_lasfile, does not work proper yet. But we might not even need to do this, since I just reported the bounding box in the report using CloudCompare. Within the main.cpp file we can work further with the filterPts.